Move tests to respective modules
