+++ connectors/filesystem/src/scanner.rs
@@ -250,3 +250,201 @@ impl FileSystemScanner {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::FileSystemScanner;
+    use crate::models::{FileSystemFile, FileSystemPermissions, FileSystemSource};
+    use std::path::PathBuf;
+    use tempfile::TempDir;
+
+    fn create_text_file(dir: &TempDir, name: &str, content: &str) -> PathBuf {
+        let path = dir.path().join(name);
+        std::fs::write(&path, content).unwrap();
+        path
+    }
+
+    fn create_docx_file(dir: &TempDir, name: &str, text: &str) -> PathBuf {
+        let path = dir.path().join(name);
+        let docx = docx_rs::Docx::new()
+            .add_paragraph(docx_rs::Paragraph::new().add_run(docx_rs::Run::new().add_text(text)));
+        let file = std::fs::File::create(&path).unwrap();
+        docx.build().pack(file).unwrap();
+        path
+    }
+
+    fn make_source(dir: &TempDir) -> FileSystemSource {
+        FileSystemSource {
+            id: "test-source".to_string(),
+            name: "Test Source".to_string(),
+            base_path: dir.path().to_path_buf(),
+            scan_interval_seconds: 300,
+            file_extensions: None,
+            exclude_patterns: None,
+            max_file_size_bytes: None,
+        }
+    }
+
+    fn make_file(path: PathBuf, mime_type: &str) -> FileSystemFile {
+        let metadata = std::fs::metadata(&path).unwrap();
+        FileSystemFile {
+            name: path.file_name().unwrap().to_string_lossy().to_string(),
+            size: metadata.len(),
+            mime_type: mime_type.to_string(),
+            created_time: metadata.created().ok(),
+            modified_time: metadata.modified().ok(),
+            is_directory: false,
+            permissions: FileSystemPermissions {
+                readable: true,
+                writable: true,
+                executable: false,
+            },
+            path,
+        }
+    }
+
+    #[tokio::test]
+    async fn test_scan_directory_discovers_files() {
+        let dir = TempDir::new().unwrap();
+        create_text_file(&dir, "notes.txt", "some text");
+        create_text_file(&dir, "readme.md", "# Hello");
+        create_docx_file(&dir, "doc.docx", "word content");
+
+        let source = make_source(&dir);
+        let scanner = FileSystemScanner::new(source);
+        let files = scanner.scan_directory().await.unwrap();
+
+        assert_eq!(files.len(), 3);
+    }
+
+    #[tokio::test]
+    async fn test_scan_directory_extension_filter() {
+        let dir = TempDir::new().unwrap();
+        create_text_file(&dir, "notes.txt", "text content");
+        create_text_file(&dir, "readme.md", "markdown content");
+        create_text_file(&dir, "code.rs", "fn main() {}");
+
+        let mut source = make_source(&dir);
+        source.file_extensions = Some(vec!["txt".to_string()]);
+
+        let scanner = FileSystemScanner::new(source);
+        let files = scanner.scan_directory().await.unwrap();
+
+        assert_eq!(files.len(), 1);
+        assert!(files[0].name.ends_with(".txt"));
+    }
+
+    #[tokio::test]
+    async fn test_scan_directory_exclude_patterns() {
+        let dir = TempDir::new().unwrap();
+        create_text_file(&dir, "notes.txt", "text content");
+
+        // Create a subdirectory with a file
+        let sub = dir.path().join("hidden");
+        std::fs::create_dir(&sub).unwrap();
+        std::fs::write(sub.join("secret.txt"), "secret").unwrap();
+
+        let mut source = make_source(&dir);
+        source.exclude_patterns = Some(vec!["hidden".to_string()]);
+
+        let scanner = FileSystemScanner::new(source);
+        let files = scanner.scan_directory().await.unwrap();
+
+        assert_eq!(files.len(), 1);
+        assert_eq!(files[0].name, "notes.txt");
+    }
+
+    #[tokio::test]
+    async fn test_scan_directory_size_limit() {
+        let dir = TempDir::new().unwrap();
+        create_text_file(&dir, "small.txt", "small");
+
+        let big_path = dir.path().join("big.txt");
+        std::fs::write(&big_path, vec![b'x'; 2000]).unwrap();
+
+        let mut source = make_source(&dir);
+        source.max_file_size_bytes = Some(1000);
+
+        let scanner = FileSystemScanner::new(source);
+        let files = scanner.scan_directory().await.unwrap();
+
+        assert_eq!(files.len(), 1);
+        assert_eq!(files[0].name, "small.txt");
+    }
+
+    #[tokio::test]
+    async fn test_read_file_content_text() {
+        let dir = TempDir::new().unwrap();
+        let path = create_text_file(&dir, "notes.txt", "Hello from scanner");
+        let file = make_file(path, "text/plain");
+
+        let source = make_source(&dir);
+        let scanner = FileSystemScanner::new(source);
+        let content = scanner.read_file_content(&file).await.unwrap();
+
+        assert_eq!(content, "Hello from scanner");
+    }
+
+    #[tokio::test]
+    async fn test_read_file_content_docx() {
+        let dir = TempDir::new().unwrap();
+        let path = create_docx_file(&dir, "doc.docx", "Scanner DOCX test");
+        let file = make_file(
+            path,
+            "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
+        );
+
+        let source = make_source(&dir);
+        let scanner = FileSystemScanner::new(source);
+        let content = scanner.read_file_content(&file).await.unwrap();
+
+        assert!(
+            content.contains("Scanner DOCX test"),
+            "Expected DOCX content via scanner, got: '{}'",
+            content
+        );
+    }
+
+    #[tokio::test]
+    async fn test_read_file_content_source_code_returns_empty() {
+        let dir = TempDir::new().unwrap();
+        let path = create_text_file(&dir, "main.rs", "fn main() {}");
+        let file = make_file(path, "text/plain");
+
+        let source = make_source(&dir);
+        let scanner = FileSystemScanner::new(source);
+        let content = scanner.read_file_content(&file).await.unwrap();
+
+        assert!(
+            content.is_empty(),
+            "Expected empty for source code, got: '{}'",
+            content
+        );
+    }
+
+    #[tokio::test]
+    async fn test_read_file_content_directory_returns_empty() {
+        let dir = TempDir::new().unwrap();
+
+        let file = FileSystemFile {
+            path: dir.path().to_path_buf(),
+            name: "testdir".to_string(),
+            size: 0,
+            mime_type: "inode/directory".to_string(),
+            created_time: None,
+            modified_time: None,
+            is_directory: true,
+            permissions: FileSystemPermissions {
+                readable: true,
+                writable: true,
+                executable: true,
+            },
+        };
+
+        let source = make_source(&dir);
+        let scanner = FileSystemScanner::new(source);
+        let content = scanner.read_file_content(&file).await.unwrap();
+
+        assert!(content.is_empty());
+    }
+}
